+++ backend/src/main/java/com/ssafy/db/repository/HashtagRepository.java
@@ -1,7 +1,13 @@
 package com.ssafy.db.repository;
 
+import com.ssafy.db.entity.Article;
 import com.ssafy.db.entity.Hashtag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HashtagRepository extends JpaRepository<Hashtag, Integer> {
+import java.util.List;
+import java.util.Optional;
+
+public interface HashtagRepository extends JpaRepository<Hashtag, Long> {
+
+//    List<Hashtag> findHashtagByArticle(Article article);
 }
